requirements.txt
streamlit
tensorflow-CPU
pillow
numpy
opencv-python-headless
exifread